# EspecificacionAPI/src/main/resources/openapi.yml
openapi: 3.0.2
info:
  title: Juegos Olímpicos
  description: "API para anadir, editar y modficar sedes de juegos olímpicos"
  
  contact:
    email: manuelpazpintor@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1-SNAPSHOT
externalDocs:
  description: Find out more about Reflectoring
  url: https://reflectoring.io/about/
servers:
- url: https://reflectoring.swagger.io/v2
tags:
- name: sedes
  description: Operaciones sobre sedes
- name: juegos
  description: Operaciones sobre juegos
 
paths:
  /juegos:
    get:
      tags:
      - juegos
      summary: Lista de sedes 
      description: Listar las ciudades que fueron  sede 
      operationId: obtenerSedes
      security:
        - basicAuth: [user] 
     
      responses:
        200:
          description: successful operation
          content:
              application/json:
                schema:
                 $ref: '#/components/schemas/JuegosArray'
        404:
          description: bad request
      x-codegen-request-body-name: body
  /juegos/buscar:
    get:
      security:
        - basicAuth: [user] 
      tags:
      - juegos
      summary: Buscar los anos que una ciudad fue sede 
      description: Anos en que una ciudad fue sede y el tipo de sede que fue
      operationId: buscarSedes
      parameters:
      - name: ciudad
        in: query
        description: 'Id de la ciudad. '
        required: true
        schema:
          type: integer
      - name: tipo
        in: query
        description: 'Tipo de sede: invierno o verano.'
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
              application/json:
              
                schema:
                 
                 $ref: '#/components/schemas/SedesArray'
        404:
          description: bad request
      x-codegen-request-body-name: body
      
  /juegos/anadir/multiple:
    post:
      security:
        - basicAuth: [user] 
      tags:
        - juegos
      summary: Anadir sedes
      operationId: anadirMultiplesSedes
     
      requestBody:
        required: true
        content:
          application/json:
            schema:
              title: modeloAnadir
              $ref: '#/components/schemas/NuevoJuegoArray'
              
                  

      responses:
        200:
          description: successful operation
          
        400:
          description: Atributos de sede no válidos
          content: {}
        404:
          description: Sede no encontrada
          content: {}
  /juegos/anadir:
    post:
      security:
        - basicAuth: [user] 
      tags:
        - juegos
      summary: Anadir sedes
      operationId: anadirSedes
     
      requestBody:
        required: true
        content:
          application/json:
            schema:
              title: modeloAnadir
              type: object
              required:
              - ano
              - tipo
              properties:
                  ano:
                    type: integer
                  tipo:
                    type: string
                  ciudad:
                    type: string
                  pais:
                     type: string
                  codigoPais:
                    type: string
                  valorPais:
                    type: integer
                  valorCiudad:
                    type: integer
                  

      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JuegosArray'
        400:
          description: Atributos de sede no válidos
          content: {}
        404:
          description: Sede no encontrada
          content: {}
  /juegos/filtrar:
    post:
      security:
        - basicAuth: [user] 
      tags:
        - juegos
      summary: Filtrar sedes
      operationId: filtrarSedes
     
      requestBody:
        required: true
        content:
          application/json:
            schema:
              title: modeloFiltrar
              type: object
              required:
              - filtro
              properties:
                  
                filtro:
                  type: string
                  
                  

      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JuegosArray'
        400:
          description: Atributos de sede no válidos
          content: {}
        404:
          description: Sede no encontrada
          content: {}
               
  /juegos/modificar:
    post:
      security:
        - basicAuth: [user] 
      tags:
        - juegos
      summary: Modificar sedes
      operationId: modificarSedes
     
      requestBody:
        required: true
        
        content:
          application/json:
            schema:
              type: object
              title: modeloEditar
              required:
              - ano
              - tipo
              properties:
                  ano:
                    type: integer
                  tipo:
                    type: string
                  ciudad:
                    type: string
                  pais:
                     type: string
                  codigoPais:
                    type: string
                  valorPais:
                    type: integer
                  idCiudad:
                    type: integer
                  nuevoAno:
                    type: integer
                  nuevoTipo:
                    type: string
                  

      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JuegosArray'
        400:
          description: Atributos de sede no válidos
          content: {}
        404:
          description: Sede no encontrada
          content: {}
          
  /juegos/eliminar:
    post:
      security:
        - basicAuth: [user] 
      tags:
        - juegos
      summary: Eliminar sedes
      operationId: eliminarSedes
     
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: modeloEliminar
              required:
              - ano
              - tipo
              properties:
                  ano:
                    type: integer
                  tipo:
                    type: string
                  

      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JuegosArray'
        400:
          description: Atributos de sede no válidos
          content: {}
        404:
          description: Sede no encontrada
          content: {}
  /juegos/login:
    post:
      security:
        - basicAuth: [user] 
      tags:
        - juegos,login
      summary: LogIn
      operationId: login
     
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: modeloLogin
              required:
              - user
              - password
              properties:
                  user:
                    type: string
                  password:
                    type: string
                  

      responses:
        200:
          description: successful operation
          content: {}
        400:
          description: Log in incorrecto
          content: {}
        404:
          description: Not found
          content: {}

 
components:
  schemas:
    Juegos:
      type: object
      properties:
        nombre_ciudad:
          type: string
        nombre_pais:
          type: string
        id_pais:
          type: integer
          format: int64
        id_ciudad:
          type: integer
          format: int64
        valor:
          type: integer
          format: int64
        descripcion_tipo_jjoo:
          type: string
        numero_veces_sede:
          type: integer
          format: int64
          
    NuevoJuego:
      type: object
      properties:
        ciudad:
          type: string
        tipo:
          type: string
        ano:
          type: integer
          format: int64
        
    Sede:
      type: object
      properties:
        descripcion_tipo:
          type: string
        ano:
         type: integer
          
          
    SedesArray:
      type: array
      items:
        $ref: '#/components/schemas/Sede'
    JuegosArray:
      type: array
      items:
        $ref: '#/components/schemas/Juegos'
        
    NuevoJuegoArray:
      type: array
      items:
        $ref: '#/components/schemas/NuevoJuego'

  securitySchemes:
    basicAuth:
        type: http
        scheme: basic
        
security:
  -    basicAuth: []
      